requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo==4.6.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Image processing
Pillow==10.1.0

# HTTP client for testing
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
bcrypt==4.1.2
cryptography==41.0.8

# QR Code generation
qrcode[pil]==7.4.2

# File handling
aiofiles==23.2.1

# Background tasks
celery==5.3.4
redis==5.0.1

# Monitoring and logging
structlog==23.2.0

# CORS
fastapi-cors==0.0.6
